Disable cli and update model
The Smite API changed to where now it can retrieve future MOTDs.
Prior, the first MOTD received would be the current MOTD, and the
rest would be the ones in the past. Now, it returns 4 MOTDs in the
future.

Because of this, this crate will now return all of the MOTDs it receives
and not just the first one.